.claude/skills/ingest/scripts/requirements.txt
# Ingest skill dependencies
pymupdf              # PDF page extraction (import fitz)
Pillow               # Image processing for OCR
